api/requirements.txt
fastapi==0.65.2
uvicorn==0.14.0
pymongo==3.5.1
python-dotenv